Allow specifying mutual exlusive flags
Warning about flags that seem mutual exclusive prevents using default
flags in Shell aliases which later could be overriden.

When parsing the command line options, the last given option for a
setting (think "mode") or (light vs. dark) counts.

Since colorls errors out when `--tree` is used together with `--all`, the
`--all` option is relaxed to `--almost-all`.

This is exactly how it works for `ls` and e.g. the famous curl tool.

Related to #103, in order to make colorls more compatible to `ls`.