Handle git status for `.` and `..` properly
The git status is only computed for the given directory and is stored as
a Hash mapping relative paths to their mode string.

For a directory, the git status is comprised of all the statuses of its files
itself and all its sub directories.

The current directory's prefix is just an empty string matching all files inside
the `@git_status` Hash.

Asking for the status of the parent directory is pointless, we do not compute it
(yet).