travis: Fix deployment condition
When building a tag on travis, the tag name is used as the branch.

By default, only the `master` branch triggers a deployment, so for a version
tag the deployment to rubygems will be skipped.

This commit fixes it, by directly matching on the `TRAVIS_BRANCH` and
`TRAVIS_TAG` variables.