Align number of link column in long listing
This fixes the misalignment of files in a long listing when some files have a
different number of characters in the links column.

Instead of iterating three times over the @contents array, in order to compute
the sizes of the variant columns, use a single loop to compute the maximum size
for each column in a one pass.