Correct git status for directories without changes
git does not report any status for empty directories, and it also does not
report any status for directories without any modified, untracked or ignored
files.

To correctly show the status, we need to check whether the directory is empty if
git did not report any status codes. When empty, show it as "not interesting",
otherwise the "no changes" green check mark is shown.

Note, I am specifically not using the untracked marker, as that would mean there
is some interesting file inside the directory which could be added.

Fixes #334.