Enable all newly introduced cops automatically
When Rubocop introduces new rules, existing code may violate them unknowingly.

That's why they are in a "pending" state initially and need to be enabled
explicitly:

```
The following cops were added to RuboCop, but are not configured. Please set
Enabled to either `true` or `false` in your `.rubocop.yml` file:

 - Layout/SpaceAroundMethodCallOperator (0.82)

 - Style/ExponentialNotation (0.82)

For more information: https://docs.rubocop.org/en/latest/versioning/
```

Since we never automatically let gem upgrade rubocop in this project, it is
safe to enable all pending cops globally. This way new cops are not silenced and
one has to deal with them when upgrading the dependency on Rubocop.