Remove man page and zsh completion file
The files are automatically generated, but using rake's `file` task does not
work correctly when the files are checked into the repo since the timestamps are
updated independently of the commit time.

Removing them ensures that these files are always freshly generated before
building the gem and thus are always up-to-date which was often not the case in
the past.