Pass a null device instead of explicitly closing stderr
When calling git any error messages should be suppressed (e.g. if an directory
is not inside a git repository).

Closing the stderr output when calling `IO.popen` works fine on POSIX systems,
but on Windows it does not and spits out a warning for every invocation:

```
git.rb:44: warning: cannot close fd before spawn
```

Fixes #517.